+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.1.1
+Update to light-baseimage:0.2.2
+
 ## 1.1.0
 Update to light-baseimage:0.2.1 :
   - Use \*.yaml.startup environment files to keep configuration secrets